requirements.txt
Django
djangorestframework
gunicorn
python-decouple
dj-database-url
dj-static
psycopg2
django-cors-headers